LRCI-3931 Move the initial build parameters into the BaseJobEntity
